+++ project/src/components/app/app.tsx
@@ -5,7 +5,8 @@ import { MainScreen } from '../../pages/main-screen/main-screen';
 import { FavoritesScreen } from '../../pages/favorites-screen/favorites-screen';
 import { getComments } from '../../mock/comments';
 import { PropertyScreen } from '../../pages/property-screen/property-screen';
-import { City, Hotel } from '../../types/hotel';
+import { Hotel } from '../../types/hotel';
+import { City } from '../../types/city';
 import { User } from '../../types/user';
 import { NotFoundScreen } from '../../pages/not-found-screen/not-found-screen';
 import { PrivateRoute } from '../private-route/pravate-route';
