+++ project/src/pages/property-screen/property-screen.tsx
@@ -61,7 +61,7 @@ export const PropertyScreen: React.FunctionComponent<PropertyScreenProps> = ({ u
                 <button className={`property__bookmark-button button ${hotel.isFavorite
                   ? 'property__bookmark-button--active'
                   : ''
-                  }`} type="button"
+                }`} type="button"
                 >
                   <svg className="property__bookmark-icon place-card__bookmark-icon" width="31" height="33">
                     <use xlinkHref="#icon-bookmark"></use>
