+++ project/src/types/city.ts
@@ -1,0 +1,6 @@
+import { Location } from './location';
+
+export type City = {
+  location: Location;
+  name: string;
+}
