+++ project/src/types/host.ts
@@ -1,0 +1,6 @@
+export type Host = {
+  avatarUrl: string;
+  id: number;
+  isPro: boolean;
+  name: string;
+}
